requirements.txt
addict==2.4.0
aniso8601==9.0.1
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
click==8.0.1
conda==4.10.1
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.24
facenet-pytorch==2.5.2
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==2.0
graphviz==0.8.4
importlib-metadata==4.8.1
intel-openmp==2021.2.0
itsdangerous==2.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mkl==2021.2.0
mxnet-cu110==1.8.0.post0
mxnet-cu112==1.8.0.post0
natsort==7.1.1
numpy==1.20.2
onnxruntime==1.4.0
onnxruntime-gpu==1.8.1
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.2
polygon==0.0.1
Polygon3==3.0.9.1
protobuf==3.17.3
pyclipper==1.3.0
pycosat==0.6.3
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
Shapely==1.7.1
tbb==2021.2.0
torch==1.9.0
torchvision==0.10.0

typing-extensions==3.10.0.2

web.py==0.40.dev0
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.1
zipp==3.5.0
PyYAML>=5.3.1
